+++ detail.html
@@ -20,6 +20,6 @@
       </div>
       <div class="review-list"></div>
     </div>
-    <script type="module" src="./src/main/detail-main.js" defer></script>
+    <script type="module" src="./src/detail/detail-main.js" defer></script>
   </body>
 </html>
