+++ src/main/main.js
@@ -1,9 +1,9 @@
-import { renderMovies } from '../data/render-movies.js';
-import { sort } from '../util/sort.js';
-import { search } from '../util/serch.js';
-import { onKeyUp } from '../util/on-key-up.js';
+import { renderMovies } from './data/render-movies.js';
+import { sort } from './util/sort.js';
+import { search } from './util/serch.js';
+import { onKeyUp } from './util/on-key-up.js';
 // import { renderMovieDetails } from "./get-movie-details.js";
-import { getId } from '../data/get-id.js';
+import { getId } from './data/get-id.js';
 
 renderMovies(); // 동작
 // Uncaught ReferenceError: search is not defined
